chore: -  move share functions to same directory -  Overwrites response 'json' to the fetcher return type
