+++ .npmignore
@@ -3,6 +3,8 @@ test
 babel.config.js
 jest.config.js
 index.js.map
+shared/index.js.map
+vanilla.js.map
 .prettierrc
 src
 .gitignore
